+++ packages/typegen/tests/__snapshots__/zod-layout-client-customer.snap.ts
@@ -1,0 +1,10 @@
+/**
+ * Put your custom overrides or transformations here.
+ * Changes to this file will NOT be overwritten.
+ */
+import { z } from "zod";
+import { Zcustomer as Zcustomer_generated } from "./generated/customer";
+
+export const Zcustomer = Zcustomer_generated;
+
+export type Tcustomer = z.infer<typeof Zcustomer>;