+++ packages/typegen/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @proofkit/typegen
 
+## 1.0.9
+
+### Patch Changes
+
+- 2ff4cd1: Update how portal validation should be passed to the fmdapi client.
+  To update, simply re-run the `npx @proofkit/typegen@latest` command and your files will be updated to the correct syntax. If you still see errors, try with the "--reset-overrides" flag to also re-create your overrides files.
+- Updated dependencies [2ff4cd1]
+  - @proofkit/fmdapi@5.0.1
+
 ## 1.0.8
 
 ### Patch Changes